+++ src/SmartHopper.Infrastructure/AIProviders/ProviderManager.cs
@@ -21,7 +21,6 @@ using System.Security.Cryptography.X509Certificates;
 using System.Threading.Tasks;
 using Rhino;
 using SmartHopper.Infrastructure.Dialogs;
-using SmartHopper.Infrastructure.Interfaces;
 using SmartHopper.Infrastructure.Settings;
 
 namespace SmartHopper.Infrastructure.AIProviders