+++ src/SmartHopper.Infrastructure/Settings/SmartHopperSettings.cs
@@ -18,8 +18,6 @@ using System.Security.Cryptography;
 using System.Text;
 using Newtonsoft.Json;
 using SmartHopper.Infrastructure.AIProviders;
-using SmartHopper.Infrastructure.AIProviders.Manager;
-using SmartHopper.Infrastructure.Models;
 
 namespace SmartHopper.Infrastructure.Settings
 {